+++ tests/input/regression/import_twice.py
@@ -1,0 +1,6 @@
+from conftest import conftest_fixture_func
+
+
+def test_using_conftest_fixture_attr():
+    if True:
+        from conftest import conftest_fixture_func