+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Fix fixtures defined with `@pytest.yield_fixture` decorator still showing FP
+- Fix crashes when using fixture + if + inline import
 
 ## [0.1.1] - 2020-05-19
 ### Fixed